# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sample/upload-image": {
      "post": {
        "operationId": "SampleRouter_uploadImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleUploadImageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadImageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Upload sample image",
        "tags": ["sample"]
      }
    },
    "/api/sample/send-message": {
      "post": {
        "operationId": "SampleRouter_sendMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleSendMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleSendMessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Send sample message to queue",
        "tags": ["sample"]
      }
    },
...
  },
  "info": {
    "title": "Gaeldle API",
    "description": "The Gaeldle API specification",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SampleUploadImageDto": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "description": "Base64 image string"
          },
          "extension": {
            "type": "string",
            "default": "jpg"
          }
        },
        "required": ["image"]
      },
      "UploadImageResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "url": {
            "type": "string"
          }
        },
        "required": ["success", "url"]
      },
      "SampleSendMessageDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"]
      },
      "SampleSendMessageResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "messageId": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["success", "message"]
      },
...
    }
  }
}